Add consul_session type
This is a first attempt at a custom type to create and
destroy consul_session resources.

Updates to resources are not supported at this time, as there
is no API call to update a session, and create/destroy will change
the ID and potentially cause issues for locks that are tied to that
ID, leading to pointless master-swaps.