Facter will only try to lock if the key is not locked
Changed to facter to do a getLockSession and only try to lock (createKV) if the
key is not locked. This will reduce the number of PUT Operations to consul as
PUT can only be responded by consul leader where GET can be responded by any
consul server with appropriate configuration.
Also added inline comments.